requirements.txt
pandas
numpy==1.26.3
pandasai
python-dotenv
openai
langchain
streamlit
notebook
ipywidgets
tqdm